requirements.txt
aiohttp==3.9.3
requests==2.31.0
tiktoken==0.6.0
streamlit==1.32.0
Pillow==10.2.0
pandas==2.2.1
openpyxl==3.1.2
chromadb==0.4.24
sentence-transformers==2.5.1
aiohttp==3.9.3
python-dotenv==1.0.1
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4 